# backend/swagger.yaml
openapi: 3.0.0
info:
  title: User Management
  description: |-
    A Server-side API built on NodeJS and MongoDB for User Management.

    Features of the APP
    - Registration of users
    - Login 
    - Reset Password
    - Update user details
    - Display list of all users
    - All the users details are stored in database with encryption

    [Github Link](https://github.com/NoahNoahkimtai)
  contact:
    email: noahkimtai24@gmail.com
  version: "1.0"
paths:
...
  /api/v1/user/update-user/{email}:
    put:
      tags:
        - User
      summary: Update User
      parameters:
        - name: email
          in: path
          required: true
          description: Email of the user to get
          schema:
            type: string
            example: user@xyz.com
      operationId: UpdateUser
      security:
        - bearer: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - name
                - phone
              type: object
              properties:
                name:
                  type: string
                  example: John Doe
                phone:
                  type: integer
                  example: 7903464439
      responses:
        "200":
          description: User updated successfully
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...